refactor(config): extract config-validator.ts (R4)
